NATIVE-397: Tidy up
