+++ src/modules/app/createAppContainer.js
@@ -5,12 +5,12 @@ import DashboardContainer from '../../routes/dashboard/containers/DashboardConta
 import CategoriesContainer from '../../routes/categories/containers/CategoriesContainer'
 import type {
   HeaderProps,
-  NavigationComponent,
+  NavigationComponent, NavigationContainer,
   NavigationNavigator,
   NavigationRouteConfig,
-  NavigationRouteConfigMap
+  NavigationRouteConfigMap, NavigationRouter, NavigationScreenProp
 } from 'react-navigation'
-import { createStackNavigator, createSwitchNavigator } from 'react-navigation'
+import { createStackNavigator, createSwitchNavigator, createAppContainer } from 'react-navigation'
 import TransparentHeaderContainer from '../layout/containers/TransparentHeaderContainer'
 import SettingsHeaderContainer from '../layout/containers/SettingsHeaderContainer'
 import HeaderContainer from '../layout/containers/HeaderContainer'
@@ -100,13 +100,33 @@ const createCityContentNavigator = (params: CreateNavigationContainerParamsType)
   }
 }
 
-const createAppNavigator = (params: CreateNavigationContainerParamsType): NavigationNavigator<*, *, *> => {
+type NavigatorPropsType = {
+  navigation: NavigationScreenProp<*>
+}
+
+const createSwitchNavigatorWithSnackbar = (
+  params: CreateNavigationContainerParamsType
+): NavigationNavigator<*, *, *> => {
   const cityContentNavigator = createCityContentNavigator(params)
-  return createSwitchNavigator({
+  const SwitchNavigator = createSwitchNavigator({
     'Intro': IntroContainer,
     'Landing': LandingContainer,
     'CityContent': cityContentNavigator
   }, { initialRouteName: params.initialRouteName })
+
+  class SwitchNavigatorWithSnackbar extends React.Component<NavigatorPropsType> {
+    static router: NavigationRouter<*, *> = SwitchNavigator.router
+    static navigationOptions = SwitchNavigator.navigationOptions
+
+    render () {
+      return <SwitchNavigator {...this.props} />
+    }
+  }
+
+  return SwitchNavigatorWithSnackbar
 }
 
-export default createAppNavigator
+const createContainer = (params: CreateNavigationContainerParamsType): NavigationContainer<*, *, *> =>
+  createAppContainer(createSwitchNavigatorWithSnackbar(params))
+
+export default createContainer
